+++ CHANGELOG.md
@@ -8,4 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`a5ba52c...main`][a5ba52c...main].
 
+### Added
+
+- Added `ReferenceToken` as a value object ([#1]), by [@localheinz]
+
 [a5ba52c...main]: https://github.com/ergebnis/json-pointer/compare/a5ba52c...main
+
+[#1]: https://github.com/ergebnis/json-pointer/pull/1
+
+[@localheinz]: https://github.com/localheinz